# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Наша CRM",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/group/{group_id}": {
            "get": {
                "tags": [
                    "Группы"
                ],
                "summary": "Получить группу по id",
                "description": "Вернет группу по id",
                "operationId": "a3cb9250261451d7f83108f0aa4ca377",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Площадки",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "test"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "test"
                                        },
                                        "area": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/group/area/{area_id}": {
            "get": {
                "tags": [
                    "Группы"
                ],
                "summary": "Получить список  групп пользователя по площадке",
                "description": "Вернет список групп текущего пользователя по площадке",
                "operationId": "0c6ac8bce42dfd02ed77d0cd1c256853",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "categories": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": ""
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": ""
                                                    },
                                                    "area": {
                                                        "type": "object",
                                                        "example": ""
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": 1,
                                                    "name": "Fanger",
                                                    "description": "Fanger",
                                                    "area": []
                                                },
                                                {
                                                    "id": 2,
                                                    "name": "Fanger",
                                                    "description": "Fanger",
                                                    "area": []
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/group/add": {
            "post": {
                "tags": [
                    "Группы"
                ],
                "summary": "Добавить группу",
                "description": "Нужно передать поля name,area_id",
                "operationId": "d78fce09fe4b79ca370f4e6155f73e2a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "area_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "test"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "test"
                                    },
                                    "area_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Вернет ошибку валидации, если поля не валидны или какие-то не отправлены"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/group/update/{group_id}": {
            "put": {
                "tags": [
                    "Группы"
                ],
                "summary": "обновить группу",
                "description": "Нужно передать поля name,city, street, house",
                "operationId": "85d44d13abae8fe538469f0ad2e72a06",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "test"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "test"
                                    },
                                    "area_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "status": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Вернет ошибку валидации, если поля не валидны или какие-то не отправлены"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/group/delete/{area_id}": {
            "delete": {
                "tags": [
                    "Группы"
                ],
                "summary": "Удалить группу по id",
                "description": "Вернет JsonResponse",
                "operationId": "bd0b4f0725136b7adf1918c6376f9511",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID группы",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    }
}